more work on margin computation
fixed the errors that occured after the last commit. now the margin
computation bare bones seem to work, but need further testing, and
integrating into the program.

But to not makes things too good, an error appears now when pressing the
start button, which seems to be caused by the event dispatch thread. It
seems to not influence the program, but I will investigate.