removed the "transitive" keyword again
My Eclipse (tested with 3 different eclipse versions) would not autocomplete imports when the "transitive" keyword was used in the module-info.
While earlier tests showed that the some parts of BEAST claim to need it, so far it still works without it.
I therefore only commented it out, showing for the future which lines would have to be replaced.
On my laptop, the project still worked with the transitive keyword. Will test again when the next java version is released