fixed missing saveAs menu item
The missing menu item to saveAs a property was added.
It was also discussed, if one menu would be missing, namely "new property".
As the user can create a new property by using the button at the property list, I think it would be best to just leave the functionality there, to not overload the UI