found the problem with windows
Everytime it runs a test it tries to find vscmd. which is just in the
beast/windows folder. VSCmd in there is just a file with a
hardcoded path where it expects the "real" vsCmd

So nothing gets set at startup, it just hopes the file is at the right
place.