started work on refactoring the BEAST api and test running
Started work on completely decoupling BEAST and the GUI (as far as
possible). I am currently experimenting with the Architecture. To this
end, I introduced a few new classes in the new package "api".
These will run the tests and communicate the results without
having to know about fxml concepts such as treelistitems.